Add and Change files for user authentication
Additions and changes are modifying the rails api generator to work with user authentication.

Application controller changed back to base to utilize it's full functionality.

The middleware for cookies, and session needed to be added back in to work with user authentication.

cookie_serializer, and session_store configured for HTTP requests